update with new speciality layout
